pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "nl2sql"
version = "0.1.0"
description = "Adaptive Hybrid NL2SQL system targeting BIRD benchmark"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "sentence-transformers>=2.7",
    "faiss-cpu>=1.8",
    "datasketch>=1.6",
    "spacy>=3.7",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
    "pandas>=2.0",
    "tqdm>=4.0",
    "tenacity>=8.0",
    "tiktoken>=0.6",
]

[project.optional-dependencies]
gemini = [
    "google-genai>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "live: marks tests that make real API calls (deselect with -m 'not live')",
]
